feat(FN-3567): review markups
